Rename 'PersonApp.py' to 'person_app.py'
